Takes the information form the user recordings
Fixed a couple of constructors that had the wrong name.  The ContainerViewModel now has a UserRecordingsDataModel property that currently loads dummy data into the user recordings... In the refresh() method, it now compares to see if there are glucose and carb recordings for certain time periods and populates those segments if they do.